refactor authentication in one activity and 2 fragments, refactor some graphic bugs
