+++ app/src/main/AndroidManifest.xml
@@ -63,7 +63,7 @@
             android:value="AIzaSyBf_Xgb8boYUw5_DlfcaLlEsWMpg14Lmwg" />
 
         <activity
-            android:name=".gui.authentication.LoginActivity"
+            android:name=".gui.authentication.AuthenticationActivity"
             android:label="@string/title_activity_login"
             android:configChanges="orientation"
             android:screenOrientation="portrait"/>
