--- app/src/main/java/org/gammf/collabora_android/app/gui/LoginFragment.java
@@ -1,41 +1,0 @@
-package org.gammf.collabora_android.app.gui;
-
-
-import android.os.Bundle;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentTransaction;
-import android.util.Log;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.Button;
-import android.widget.EditText;
-import android.widget.ProgressBar;
-import android.widget.TextView;
-import android.widget.Toast;
-
-import com.loopj.android.http.AsyncHttpClient;
-import com.loopj.android.http.AsyncHttpResponseHandler;
-
-import org.gammf.collabora_android.app.R;
-import org.gammf.collabora_android.utils.AuthenticationUtils;
-import org.gammf.collabora_android.utils.LocalStorageUtils;
-import org.gammf.collabora_android.utils.MandatoryFieldMissingException;
-import org.gammf.collabora_android.utils.UserUtils;
-import org.json.JSONException;
-import org.json.JSONObject;
-import org.mindrot.jbcrypt.BCrypt;
-
-import java.io.IOException;
-
-import cz.msebera.android.httpclient.Header;
-
-/**
- * A simple {@link Fragment} subclass.
- * Use the {@link LoginFragment#newInstance} factory method to
- * create an instance of this fragment.
- */
-public class LoginFragment extends Fragment {
-
-
-}
