+++ app/src/main/res/layout/fragment_registration.xml
@@ -54,9 +54,8 @@
                     android:layout_height="30dp"
                     android:hint="@string/prompt_birth"
                     android:maxLines="1"
-                    android:textSize="22sp"
+                    android:textSize="20sp"
                     android:layout_alignBottom="@+id/txtNewDateSelected"
-                    android:layout_below="@+id/surname"
                     android:layout_alignParentStart="true"
                     android:layout_toStartOf="@+id/txtNewDateSelected"/>
 
@@ -65,11 +64,10 @@
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
                     android:text="@string/dateStr"
-                    android:textSize="30sp"
+                    android:textSize="20sp"
                     android:gravity="center"
                     android:layout_alignBottom="@+id/btnSetDate"
                     android:layout_alignEnd="@+id/btnSetDate"
-                    android:layout_below="@+id/surname"
                     android:layout_marginEnd="76dp"
                     android:layout_marginBottom="9dp"/>
